AVRO-2108: [Java] DoubleArray converts double to float (#3285)
* fixing DoubleArray float cast

* add new line

* preventing boxing

---------

+++ lang/java/avro/src/main/java/org/apache/avro/generic/PrimitivesArrays.java
@@ -550,7 +550,7 @@ public class PrimitivesArrays {
  if (o == null) {
  return;
  }
- this.add(location, o.floatValue());
+ this.add(location, o.doubleValue());
  }
 
  public void add(int location, double o) {
